WAITP-1343 Reuse types
